move helper methods to helper file
